refactor: extract dive
